Refactored and moved some files
